# Improving drastically the performance of the utf-8 to utf-16 icelake transcoder in some cases