Over some inputs dominated by 1-byte and 2-byte UTF-8 characters, the icelake utf-8 to utf-16 transcoder has very low instruction/cycle count. The reason is a long dependency chain. We can break the chain at the expense of a higher instruction count.

The idea is that we load 64 bytes with each iteration, and if we detect that they are made of 1-byte or 2-byte characters, then we always decode fully the set of characters (between 16 and 32) that begin in the first 32 bytes of our input. Of course, that's pessimistic because we can often process much more if there are few ASCII (1-byte) characters.

For lipsum/Arabic, this change pushes our from ~4.7 GB/s to 7.8 GB/s under GCC and from 5.0 GB/s to 7.6 GB/s under LLVM.

My approach is simplistic and might be improved, but the gains appear quite solid.

*Warning*: the cycle/byte metric should be ignored, it is reporting an erroneous value, see https://github.com/simdutf/simdutf/pull/194

lipsum files:
```
first the new code, then the old code
We start with GCC 12
unicode_lipsum/lipsum/Arabic-Lipsum.utf8.txt
 -- new : --
   1.373 ins/byte,    1.373 cycle/byte,    7.828 GB/s (0.3 %),     3.418 GHz,    3.144 ins/cycle 
 -- old : --
   0.669 ins/byte,    0.669 cycle/byte,    4.692 GB/s (0.2 %),     3.411 GHz,    0.920 ins/cycle 
unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
 -- new : --
   1.213 ins/byte,    1.213 cycle/byte,    5.811 GB/s (0.3 %),     3.415 GHz,    2.064 ins/cycle 
 -- old : --
   1.213 ins/byte,    1.213 cycle/byte,    5.815 GB/s (0.3 %),     3.414 GHz,    2.066 ins/cycle 
unicode_lipsum/lipsum/Emoji-Lipsum.utf8.txt
 -- new : --
   1.553 ins/byte,    1.553 cycle/byte,    3.795 GB/s (0.3 %),     3.411 GHz,    1.728 ins/cycle 
 -- old : --
   1.553 ins/byte,    1.553 cycle/byte,    3.798 GB/s (0.4 %),     3.410 GHz,    1.729 ins/cycle 
unicode_lipsum/lipsum/Hebrew-Lipsum.utf8.txt
 -- new : --
   1.374 ins/byte,    1.374 cycle/byte,    7.857 GB/s (0.4 %),     3.420 GHz,    3.156 ins/cycle 
 -- old : --
   0.670 ins/byte,    0.670 cycle/byte,    4.638 GB/s (0.3 %),     3.413 GHz,    0.911 ins/cycle 
unicode_lipsum/lipsum/Hindi-Lipsum.utf8.txt
 -- new : --
   1.228 ins/byte,    1.228 cycle/byte,    4.910 GB/s (0.3 %),     3.409 GHz,    1.768 ins/cycle 
 -- old : --
   1.228 ins/byte,    1.228 cycle/byte,    4.907 GB/s (0.3 %),     3.412 GHz,    1.766 ins/cycle 
unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
 -- new : --
   1.220 ins/byte,    1.220 cycle/byte,    5.408 GB/s (0.4 %),     3.414 GHz,    1.933 ins/cycle 
 -- old : --
   1.220 ins/byte,    1.220 cycle/byte,    5.419 GB/s (0.3 %),     3.414 GHz,    1.937 ins/cycle 
unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
 -- new : --
   1.230 ins/byte,    1.230 cycle/byte,    4.873 GB/s (0.3 %),     3.415 GHz,    1.755 ins/cycle 
 -- old : --
   1.230 ins/byte,    1.230 cycle/byte,    5.757 GB/s (24.2 %),     3.415 GHz,    2.074 ins/cycle 
unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
 -- new : --
   0.253 ins/byte,    0.253 cycle/byte,   23.135 GB/s (5.0 %),     3.449 GHz,    1.698 ins/cycle 
 -- old : --
   0.253 ins/byte,    0.253 cycle/byte,   23.067 GB/s (4.6 %),     3.451 GHz,    1.692 ins/cycle 
unicode_lipsum/lipsum/Russian-Lipsum.utf8.txt
 -- new : --
   1.373 ins/byte,    1.373 cycle/byte,    7.840 GB/s (0.3 %),     3.414 GHz,    3.152 ins/cycle 
 -- old : --
   0.660 ins/byte,    0.660 cycle/byte,    5.912 GB/s (42.9 %),     3.410 GHz,    1.144 ins/cycle 
Now LLVM
unicode_lipsum/lipsum/Arabic-Lipsum.utf8.txt
 -- new : --
   1.297 ins/byte,    1.297 cycle/byte,    7.613 GB/s (0.2 %),     3.419 GHz,    2.889 ins/cycle 
 -- old : --
   0.634 ins/byte,    0.634 cycle/byte,    4.984 GB/s (0.3 %),     3.413 GHz,    0.926 ins/cycle 
unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
 -- new : --
   1.086 ins/byte,    1.086 cycle/byte,    5.354 GB/s (0.3 %),     3.416 GHz,    1.702 ins/cycle 
 -- old : --
   1.086 ins/byte,    1.086 cycle/byte,    5.355 GB/s (0.3 %),     3.418 GHz,    1.701 ins/cycle 
unicode_lipsum/lipsum/Emoji-Lipsum.utf8.txt
 -- new : --
   1.350 ins/byte,    1.350 cycle/byte,    4.174 GB/s (0.5 %),     3.413 GHz,    1.650 ins/cycle 
 -- old : --
   1.350 ins/byte,    1.350 cycle/byte,    4.173 GB/s (0.4 %),     3.413 GHz,    1.650 ins/cycle 
unicode_lipsum/lipsum/Hebrew-Lipsum.utf8.txt
 -- new : --
   1.298 ins/byte,    1.298 cycle/byte,    7.698 GB/s (0.4 %),     3.423 GHz,    2.919 ins/cycle 
 -- old : --
   0.636 ins/byte,    0.636 cycle/byte,    4.924 GB/s (0.3 %),     3.416 GHz,    0.917 ins/cycle 
unicode_lipsum/lipsum/Hindi-Lipsum.utf8.txt
 -- new : --
   1.099 ins/byte,    1.099 cycle/byte,    4.563 GB/s (0.2 %),     3.410 GHz,    1.471 ins/cycle 
 -- old : --
   1.099 ins/byte,    1.099 cycle/byte,    4.563 GB/s (0.3 %),     3.411 GHz,    1.471 ins/cycle 
unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
 -- new : --
   1.092 ins/byte,    1.092 cycle/byte,    5.000 GB/s (0.2 %),     3.416 GHz,    1.599 ins/cycle 
 -- old : --
   1.092 ins/byte,    1.092 cycle/byte,    5.001 GB/s (0.3 %),     3.415 GHz,    1.600 ins/cycle 
unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
 -- new : --
   1.101 ins/byte,    1.101 cycle/byte,    4.524 GB/s (0.2 %),     3.415 GHz,    1.459 ins/cycle 
 -- old : --
   1.101 ins/byte,    1.101 cycle/byte,    4.530 GB/s (0.2 %),     3.415 GHz,    1.461 ins/cycle 
unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
 -- new : --
   0.238 ins/byte,    0.238 cycle/byte,   24.442 GB/s (6.7 %),     3.466 GHz,    1.677 ins/cycle 
 -- old : --
   0.238 ins/byte,    0.238 cycle/byte,   24.657 GB/s (7.4 %),     3.462 GHz,    1.694 ins/cycle 
unicode_lipsum/lipsum/Russian-Lipsum.utf8.txt
 -- new : --
   1.298 ins/byte,    1.298 cycle/byte,    7.645 GB/s (0.3 %),     3.416 GHz,    2.904 ins/cycle 
 -- old : --
   0.626 ins/byte,    0.626 cycle/byte,    5.001 GB/s (0.3 %),     3.410 GHz,    0.917 ins/cycle 
```

wikipedia/Mars files:
```
first the new code, then the old code
We start with GCC 12
unicode_lipsum/wikipedia_mars/arabic.utf8.txt
 -- new : --
   1.047 ins/byte,    1.047 cycle/byte,    8.909 GB/s (1.5 %),     3.404 GHz,    2.740 ins/cycle 
 -- old : --
   0.699 ins/byte,    0.699 cycle/byte,    4.982 GB/s (0.7 %),     3.402 GHz,    1.024 ins/cycle 
unicode_lipsum/wikipedia_mars/chinese.utf8.txt
 -- new : --
   1.412 ins/byte,    1.412 cycle/byte,    4.146 GB/s (0.2 %),     3.404 GHz,    1.720 ins/cycle 
 -- old : --
   1.406 ins/byte,    1.406 cycle/byte,    4.059 GB/s (0.2 %),     3.404 GHz,    1.677 ins/cycle 
unicode_lipsum/wikipedia_mars/czech.utf8.txt
 -- new : --
   1.231 ins/byte,    1.231 cycle/byte,    7.044 GB/s (0.3 %),     3.408 GHz,    2.544 ins/cycle 
 -- old : --
   1.097 ins/byte,    1.097 cycle/byte,    3.811 GB/s (0.2 %),     3.404 GHz,    1.228 ins/cycle 
unicode_lipsum/wikipedia_mars/english.utf8.txt
 -- new : --
   0.389 ins/byte,    0.389 cycle/byte,   13.748 GB/s (2.1 %),     3.410 GHz,    1.569 ins/cycle 
 -- old : --
   0.384 ins/byte,    0.384 cycle/byte,   14.155 GB/s (1.8 %),     3.410 GHz,    1.593 ins/cycle 
unicode_lipsum/wikipedia_mars/esperanto.utf8.txt
 -- new : --
   0.896 ins/byte,    0.896 cycle/byte,    8.703 GB/s (0.7 %),     3.417 GHz,    2.282 ins/cycle 
 -- old : --
   0.843 ins/byte,    0.843 cycle/byte,    5.928 GB/s (0.6 %),     3.412 GHz,    1.464 ins/cycle 
unicode_lipsum/wikipedia_mars/french.utf8.txt
 -- new : --
   0.989 ins/byte,    0.989 cycle/byte,    9.001 GB/s (1.0 %),     3.407 GHz,    2.614 ins/cycle 
 -- old : --
   0.882 ins/byte,    0.882 cycle/byte,    4.654 GB/s (0.6 %),     3.403 GHz,    1.207 ins/cycle 
unicode_lipsum/wikipedia_mars/german.utf8.txt
 -- new : --
   0.880 ins/byte,    0.880 cycle/byte,    9.253 GB/s (0.6 %),     3.407 GHz,    2.389 ins/cycle 
 -- old : --
   0.821 ins/byte,    0.821 cycle/byte,    6.218 GB/s (0.4 %),     3.405 GHz,    1.500 ins/cycle 
unicode_lipsum/wikipedia_mars/greek.utf8.txt
 -- new : --
   1.114 ins/byte,    1.114 cycle/byte,    8.458 GB/s (0.3 %),     3.408 GHz,    2.766 ins/cycle 
 -- old : --
   0.776 ins/byte,    0.776 cycle/byte,    4.878 GB/s (0.3 %),     3.405 GHz,    1.112 ins/cycle 
unicode_lipsum/wikipedia_mars/hebrew.utf8.txt
 -- new : --
   1.242 ins/byte,    1.242 cycle/byte,    7.602 GB/s (0.3 %),     3.407 GHz,    2.771 ins/cycle 
 -- old : --
   0.878 ins/byte,    0.878 cycle/byte,    4.085 GB/s (0.2 %),     3.404 GHz,    1.054 ins/cycle 
unicode_lipsum/wikipedia_mars/hindi.utf8.txt
 -- new : --
   1.126 ins/byte,    1.126 cycle/byte,    5.329 GB/s (0.3 %),     3.402 GHz,    1.764 ins/cycle 
 -- old : --
   1.119 ins/byte,    1.119 cycle/byte,    5.267 GB/s (0.3 %),     3.403 GHz,    1.732 ins/cycle 
unicode_lipsum/wikipedia_mars/japanese.utf8.txt
 -- new : --
   1.346 ins/byte,    1.346 cycle/byte,    4.394 GB/s (0.3 %),     3.405 GHz,    1.738 ins/cycle 
 -- old : --
   1.341 ins/byte,    1.341 cycle/byte,    4.259 GB/s (0.3 %),     3.405 GHz,    1.678 ins/cycle 
unicode_lipsum/wikipedia_mars/korean.utf8.txt
 -- new : --
   1.408 ins/byte,    1.408 cycle/byte,    4.199 GB/s (0.3 %),     3.407 GHz,    1.735 ins/cycle 
 -- old : --
   1.402 ins/byte,    1.402 cycle/byte,    4.035 GB/s (0.2 %),     3.407 GHz,    1.661 ins/cycle 
unicode_lipsum/wikipedia_mars/persan.utf8.txt
 -- new : --
   1.183 ins/byte,    1.183 cycle/byte,    6.586 GB/s (0.4 %),     3.407 GHz,    2.286 ins/cycle 
 -- old : --
   0.989 ins/byte,    0.989 cycle/byte,    4.468 GB/s (0.3 %),     3.405 GHz,    1.297 ins/cycle 
unicode_lipsum/wikipedia_mars/portuguese.utf8.txt
 -- new : --
   0.970 ins/byte,    0.970 cycle/byte,    9.124 GB/s (0.4 %),     3.406 GHz,    2.598 ins/cycle 
 -- old : --
   0.885 ins/byte,    0.885 cycle/byte,    4.966 GB/s (0.3 %),     3.403 GHz,    1.291 ins/cycle 
unicode_lipsum/wikipedia_mars/russian.utf8.txt
 -- new : --
   1.155 ins/byte,    1.155 cycle/byte,    8.067 GB/s (0.4 %),     3.406 GHz,    2.736 ins/cycle 
 -- old : --
   0.799 ins/byte,    0.799 cycle/byte,    4.601 GB/s (0.4 %),     3.402 GHz,    1.080 ins/cycle 
unicode_lipsum/wikipedia_mars/thai.utf8.txt
 -- new : --
   1.036 ins/byte,    1.036 cycle/byte,    5.878 GB/s (0.7 %),     3.403 GHz,    1.789 ins/cycle 
 -- old : --
   1.032 ins/byte,    1.032 cycle/byte,    5.813 GB/s (0.8 %),     3.402 GHz,    1.764 ins/cycle 
unicode_lipsum/wikipedia_mars/turkish.utf8.txt
 -- new : --
   1.054 ins/byte,    1.054 cycle/byte,    8.461 GB/s (0.5 %),     3.407 GHz,    2.617 ins/cycle 
 -- old : --
   0.928 ins/byte,    0.928 cycle/byte,    4.582 GB/s (0.2 %),     3.404 GHz,    1.249 ins/cycle 
unicode_lipsum/wikipedia_mars/vietnamese.utf8.txt
 -- new : --
   1.554 ins/byte,    1.554 cycle/byte,    3.909 GB/s (0.3 %),     3.403 GHz,    1.785 ins/cycle 
 -- old : --
   1.543 ins/byte,    1.543 cycle/byte,    3.724 GB/s (0.3 %),     3.402 GHz,    1.689 ins/cycle 
Now LLVM
unicode_lipsum/wikipedia_mars/arabic.utf8.txt
 -- new : --
   0.987 ins/byte,    0.987 cycle/byte,    8.033 GB/s (1.3 %),     3.406 GHz,    2.327 ins/cycle 
 -- old : --
   0.657 ins/byte,    0.657 cycle/byte,    5.003 GB/s (1.4 %),     3.403 GHz,    0.966 ins/cycle 
unicode_lipsum/wikipedia_mars/chinese.utf8.txt
 -- new : --
   1.268 ins/byte,    1.268 cycle/byte,    3.773 GB/s (0.3 %),     3.405 GHz,    1.404 ins/cycle 
 -- old : --
   1.262 ins/byte,    1.262 cycle/byte,    3.722 GB/s (0.5 %),     3.406 GHz,    1.379 ins/cycle 
unicode_lipsum/wikipedia_mars/czech.utf8.txt
 -- new : --
   1.148 ins/byte,    1.148 cycle/byte,    6.535 GB/s (1.1 %),     3.409 GHz,    2.200 ins/cycle 
 -- old : --
   1.016 ins/byte,    1.016 cycle/byte,    3.896 GB/s (0.3 %),     3.406 GHz,    1.162 ins/cycle 
unicode_lipsum/wikipedia_mars/english.utf8.txt
 -- new : --
   0.360 ins/byte,    0.360 cycle/byte,   13.995 GB/s (2.2 %),     3.411 GHz,    1.476 ins/cycle 
 -- old : --
   0.354 ins/byte,    0.354 cycle/byte,   13.818 GB/s (2.2 %),     3.408 GHz,    1.437 ins/cycle 
unicode_lipsum/wikipedia_mars/esperanto.utf8.txt
 -- new : --
   0.828 ins/byte,    0.828 cycle/byte,    8.048 GB/s (1.3 %),     3.421 GHz,    1.948 ins/cycle 
 -- old : --
   0.775 ins/byte,    0.775 cycle/byte,    5.895 GB/s (0.7 %),     3.414 GHz,    1.338 ins/cycle 
unicode_lipsum/wikipedia_mars/french.utf8.txt
 -- new : --
   0.929 ins/byte,    0.929 cycle/byte,    7.498 GB/s (1.6 %),     3.406 GHz,    2.046 ins/cycle 
 -- old : --
   0.824 ins/byte,    0.824 cycle/byte,    4.427 GB/s (1.6 %),     3.403 GHz,    1.072 ins/cycle 
unicode_lipsum/wikipedia_mars/german.utf8.txt
 -- new : --
   0.814 ins/byte,    0.814 cycle/byte,    8.567 GB/s (0.8 %),     3.408 GHz,    2.045 ins/cycle 
 -- old : --
   0.756 ins/byte,    0.756 cycle/byte,    6.200 GB/s (0.5 %),     3.407 GHz,    1.376 ins/cycle 
unicode_lipsum/wikipedia_mars/greek.utf8.txt
 -- new : --
   1.044 ins/byte,    1.044 cycle/byte,    7.772 GB/s (0.7 %),     3.409 GHz,    2.381 ins/cycle 
 -- old : --
   0.724 ins/byte,    0.724 cycle/byte,    5.042 GB/s (0.4 %),     3.406 GHz,    1.072 ins/cycle 
unicode_lipsum/wikipedia_mars/hebrew.utf8.txt
 -- new : --
   1.165 ins/byte,    1.165 cycle/byte,    7.018 GB/s (0.5 %),     3.408 GHz,    2.399 ins/cycle 
 -- old : --
   0.819 ins/byte,    0.819 cycle/byte,    4.208 GB/s (0.3 %),     3.405 GHz,    1.013 ins/cycle 
unicode_lipsum/wikipedia_mars/hindi.utf8.txt
 -- new : --
   1.012 ins/byte,    1.012 cycle/byte,    4.647 GB/s (0.6 %),     3.403 GHz,    1.381 ins/cycle 
 -- old : --
   1.005 ins/byte,    1.005 cycle/byte,    4.626 GB/s (0.8 %),     3.404 GHz,    1.366 ins/cycle 
unicode_lipsum/wikipedia_mars/japanese.utf8.txt
 -- new : --
   1.209 ins/byte,    1.209 cycle/byte,    4.001 GB/s (0.4 %),     3.405 GHz,    1.421 ins/cycle 
 -- old : --
   1.204 ins/byte,    1.204 cycle/byte,    3.920 GB/s (0.6 %),     3.405 GHz,    1.386 ins/cycle 
unicode_lipsum/wikipedia_mars/korean.utf8.txt
 -- new : --
   1.264 ins/byte,    1.264 cycle/byte,    3.843 GB/s (0.6 %),     3.408 GHz,    1.426 ins/cycle 
 -- old : --
   1.259 ins/byte,    1.259 cycle/byte,    3.726 GB/s (0.5 %),     3.408 GHz,    1.376 ins/cycle 
unicode_lipsum/wikipedia_mars/persan.utf8.txt
 -- new : --
   1.091 ins/byte,    1.091 cycle/byte,    6.085 GB/s (0.5 %),     3.408 GHz,    1.948 ins/cycle 
 -- old : --
   0.907 ins/byte,    0.907 cycle/byte,    4.457 GB/s (0.4 %),     3.406 GHz,    1.187 ins/cycle 
unicode_lipsum/wikipedia_mars/portuguese.utf8.txt
 -- new : --
   0.906 ins/byte,    0.906 cycle/byte,    8.437 GB/s (0.7 %),     3.409 GHz,    2.243 ins/cycle 
 -- old : --
   0.822 ins/byte,    0.822 cycle/byte,    5.089 GB/s (0.5 %),     3.404 GHz,    1.229 ins/cycle 
unicode_lipsum/wikipedia_mars/russian.utf8.txt
 -- new : --
   1.083 ins/byte,    1.083 cycle/byte,    7.130 GB/s (0.9 %),     3.406 GHz,    2.266 ins/cycle 
 -- old : --
   0.744 ins/byte,    0.744 cycle/byte,    4.665 GB/s (0.7 %),     3.403 GHz,    1.020 ins/cycle 
unicode_lipsum/wikipedia_mars/thai.utf8.txt
 -- new : --
   0.931 ins/byte,    0.931 cycle/byte,    5.022 GB/s (0.7 %),     3.402 GHz,    1.374 ins/cycle 
 -- old : --
   0.928 ins/byte,    0.928 cycle/byte,    5.001 GB/s (0.9 %),     3.403 GHz,    1.363 ins/cycle 
unicode_lipsum/wikipedia_mars/turkish.utf8.txt
 -- new : --
   0.985 ins/byte,    0.985 cycle/byte,    7.827 GB/s (0.6 %),     3.410 GHz,    2.261 ins/cycle 
 -- old : --
   0.862 ins/byte,    0.862 cycle/byte,    4.705 GB/s (0.4 %),     3.405 GHz,    1.192 ins/cycle 
unicode_lipsum/wikipedia_mars/vietnamese.utf8.txt
 -- new : --
   1.397 ins/byte,    1.397 cycle/byte,    3.476 GB/s (0.7 %),     3.402 GHz,    1.427 ins/cycle 
 -- old : --
   1.386 ins/byte,    1.386 cycle/byte,    3.373 GB/s (0.3 %),     3.402 GHz,    1.374 ins/cycle 
```