@clausecker 2022-12-14
How does this perform with the Wikipedia-mars datasets?  I imagine the performance gains might be reversed for languages like French and German.

I had previously played with the idea of only processing 32 byte of input data for the 2-byte case, allowing us to omit the complex length computations.  However, I abandoned the idea as it did not seem to yield promising results.  Fascinating how it actually does yield good results!